Remove external databases, including SQLite
Baking the data into the binary means there's no need for external data
files. These files could be missing because they weren't installed
correctly. They're also redundant (and were inconsistent) with each
other.

This is also smaller and faster than SQLite since it doesn't involve
opening a database, then analyzing and compiling a dynamically
constructed (!) SQL expression. This is especially true since the
database contained no relevant indexes, and so SQLite was doing the same
linear table scan and sort. (Though the table is so small that this
hardly matters.)

With some cleverer database organization, even the sort could be
skipped, but it's probably more important to be readable.